Added backend checks for user auth, user membership in project when trying to update and is project archived or not

+++ LV6/projects/server/controller/ProjectController.js
@@ -2,6 +2,9 @@ const Projectdb = require('../model/project');
 
 // Retrieve and return all projects / retrieve and return a single project
 exports.find = (req,res)=>{
+    if(!req.isAuthenticated()){
+        return res.status(401).send('You are not logged in');
+    }
 
     if(req.query.id){
         const id = req.query.id;
@@ -32,14 +35,14 @@ exports.find = (req,res)=>{
 
 // Create and save a new project
 exports.create = (req,res)=>{
+    if (!req.isAuthenticated()){
+        res.status(401).send({message: 'You are not logged in!'});
+    }
+
     // validate request
     if(!req.body){
         res.status(400).send({ message: "Content can not be empty!"});
         return;
-    }
-
-    if (!req.isAuthenticated()){
-        res.status(401).send({message: 'You are not logged in!'});
     }
 
     // new project
@@ -73,6 +76,10 @@ exports.create = (req,res)=>{
 
 // Update a project by id
 exports.update = (req, res) => {
+    if(!req.isAuthenticated()){
+        return res.status(401).send('You are not logged in');
+    }
+
     if(!req.body){
         return res
             .status(400)
@@ -80,15 +87,38 @@ exports.update = (req, res) => {
     }
 
     const id = req.params.id;
-    req.body.updated_at = new Date(Date.now());
 
+    Projectdb.findById(id)
+        .then(project => {
+            if(project.finished_at){
+                // Project is archived and cannot be updated
+                return res.send(project);
+            }
+            // Update updated_at attribute
+            project.updated_at = new Date(Date.now());
+            // Only owner can update everything on project
+            if(project.owner.id === req.user._id.toString()){
+                updateProject(id, req.body, res)
+            }
+            // Members can change only tasks_done
+            else if(project.members.find(member => member._id.equals(req.user._id))){
+                project.tasks_done = req.body.tasks_done
+                updateProject(id, project, res);
+            }
+            else{
+                // Current user isn't a project member
+                res.status(401).send('Unauthorized');
+            }
+        })
+}
 
-    Projectdb.findByIdAndUpdate(id, req.body, { new: true})
-        .then(data => {
-            if(!data){
+function updateProject(id, data, res){
+    Projectdb.findByIdAndUpdate(id, data, { new: true})
+        .then(project => {
+            if(!project){
                 res.status(404).send({ message: `Cannot update project with ${id}.`})
             }else{
-                res.send(data);
+                res.send(project);
             }
         })
         .catch(err => {
@@ -98,6 +128,10 @@ exports.update = (req, res) => {
 
 // Delete a project by id
 exports.delete = (req, res) => {
+    if(!req.isAuthenticated()){
+        return res.status(401).send('You are not logged in');
+    }
+
     const id = req.params.id;
     Projectdb.findByIdAndDelete(id)
         .then(data => {
@@ -117,6 +151,10 @@ exports.delete = (req, res) => {
 }
 
 exports.archive = (req, res) => {
+    if(!req.isAuthenticated()){
+        return res.status(401).send('You are not logged in');
+    }
+
     const id = req.params.id;
 
     // Get project by id
@@ -129,17 +167,7 @@ exports.archive = (req, res) => {
             }else{
                 data.finished_at = new Date(Date.now());
                 // Update the project with archived date
-                Projectdb.findByIdAndUpdate(id, data, { new: true })
-                    .then(data => {
-                        if(!data){
-                            res.status(404).send({ message: `Cannot update project with ${id}.`})
-                        }else{
-                            res.send(data);
-                        }
-                    })
-                    .catch(err => {
-                        res.status(500).send({message: "Invalid project information"});
-                    });
+                updateProject(id, data, res);
             }
         })
         .catch(err => {
